@@ -1,13 +1,21 @@
 module github.com/matrixik/sort-camera-photos
 
-go 1.14
+go 1.21
 
 require (
-	github.com/evanoberholster/exiftool v0.0.0-20200626054632-b40ffd90f1c3
-	github.com/go-errors/errors v1.1.1
-	github.com/golang/geo v0.0.0-20200319012246-673a6f80352d // indirect
+	github.com/evanoberholster/imagemeta v0.3.1
+	github.com/go-errors/errors v1.5.1
 	github.com/google/go-cmp v0.5.0
 	github.com/gosimple/hashdir v1.0.1
 	github.com/otiai10/copy v1.2.0
-	github.com/rs/zerolog v1.19.0
+	github.com/rs/zerolog v1.32.0
+)
+
+require (
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/philhofer/fwd v1.1.2 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/tinylib/msgp v1.1.9 // indirect
+	golang.org/x/sys v0.17.0 // indirect
 )
